# docs/swagger.yaml
basePath: /
definitions:
...
  model.MstApproval:
    properties:
      action:
        type: string
      approval_users:
        items:
          $ref: '#/definitions/model.MstApprovalUser'
        type: array
      count:
        type: integer
      created_at:
        type: string
      created_by:
        $ref: '#/definitions/model.MstUser'
      id:
        type: integer
      id_createdby:
        type: integer
      id_menu:
        type: integer
      id_updatedby:
        type: integer
      level:
        type: integer
      menu:
        $ref: '#/definitions/model.MstMenu'
      status:
        type: string
      updated_at:
        type: string
      updated_by:
        $ref: '#/definitions/model.MstUser'
    type: object
  model.MstApprovalUser:
    properties:
      approval:
        $ref: '#/definitions/model.MstApproval'
      id_approval:
        type: integer
      id_user:
        type: integer
      user:
        $ref: '#/definitions/model.MstUser'
    type: object
  model.MstDept:
    properties:
      code:
        type: string
      created_at:
        type: string
      created_by:
        $ref: '#/definitions/model.MstUser'
      description:
        type: string
      id:
        type: integer
      id_createdby:
        type: integer
      id_updatedby:
        type: integer
      remarks:
        type: string
      updated_at:
        type: string
      updated_by:
        $ref: '#/definitions/model.MstUser'
    type: object
...
  model.MstMenu:
    properties:
      children:
        items:
          $ref: '#/definitions/model.MstMenu'
        type: array
      created_at:
        type: string
      created_by:
        $ref: '#/definitions/model.MstUser'
      icon:
        type: string
      id:
        type: integer
      id_createdby:
        type: integer
      id_parent:
        type: integer
      id_updatedby:
        type: integer
      is_delete:
        type: integer
      label:
        type: string
      menu_approvals:
        items:
          $ref: '#/definitions/model.MstApproval'
        type: array
      parent:
        $ref: '#/definitions/model.MstMenu'
      path:
        type: string
      role_permissions:
        items:
          $ref: '#/definitions/model.MstRolePermission'
        type: array
      sort:
        type: integer
      updated_at:
        type: string
      updated_by:
        $ref: '#/definitions/model.MstUser'
    type: object
...
  model.MstRole:
    properties:
      created_at:
        type: string
      created_by:
        $ref: '#/definitions/model.MstUser'
      id:
        type: integer
      id_createdby:
        type: integer
      id_updatedby:
        type: integer
      name:
        type: string
      role_menus:
        items:
          $ref: '#/definitions/model.MstRoleMenu'
        type: array
      role_permissions:
        items:
          $ref: '#/definitions/model.MstRolePermission'
        type: array
      updated_at:
        type: string
      updated_by:
        $ref: '#/definitions/model.MstUser'
    type: object
  model.MstRoleMenu:
    properties:
      id_menu:
        type: integer
      id_role:
        type: integer
      menu:
        $ref: '#/definitions/model.MstMenu'
      role:
        $ref: '#/definitions/model.MstRole'
    type: object
  model.MstRolePermission:
    properties:
      id_menu:
        type: integer
      id_role:
        type: integer
      is_create:
        type: boolean
      is_delete:
        type: boolean
      is_update:
        type: boolean
      menu:
        $ref: '#/definitions/model.MstMenu'
      role:
        $ref: '#/definitions/model.MstRole'
    type: object
  model.MstUser:
    properties:
      created_at:
        type: string
      created_by:
        $ref: '#/definitions/model.MstUser'
      dept:
        $ref: '#/definitions/model.MstDept'
      email:
        type: string
      id:
        type: integer
      id_createdby:
        type: integer
      id_dept:
        type: integer
      id_updatedby:
        type: integer
      is_active:
        type: boolean
      is_two_fa:
        type: boolean
      name:
        type: string
      otp_key:
        type: string
      otp_url:
        type: string
      password:
        type: string
      refresh_token:
        type: string
      updated_at:
        type: string
      updated_by:
        $ref: '#/definitions/model.MstUser'
      user_roles:
        items:
          $ref: '#/definitions/model.MstUserRole'
        type: array
      username:
        type: string
    type: object
  model.MstUserRole:
    properties:
      id_role:
        type: integer
      id_user:
        type: integer
      role:
        $ref: '#/definitions/model.MstRole'
      user:
        $ref: '#/definitions/model.MstUser'
    type: object
...
host: localhost:5050/api
info:
  contact: {}
  description: INSIST adalah sistem manajemen manufaktur milik PT. Indoseiki Metalutama.
    REST API Documentation untuk INSIST menjelaskan bagaimana pengembang dapat menggunakan
    API ini untuk mengelola dan mengakses data manufaktur perusahaan dengan aman dan
    efisien, menggunakan protokol HTTP dan metode REST seperti GET, POST, PUT, & DELETE.
  title: INSIST REST API Documentation
  version: "1.0"
paths:
...
  /admin/approval-user:
    get:
      consumes:
      - application/json
      description: Retrieves approval userss with pagination and optional search
      parameters:
      - default: 1
        description: Page number
        in: query
        name: page
        type: integer
      - default: 20
        description: Number of rows per page
        in: query
        name: rows
        type: integer
      - description: Search keyword for filtering approval users
        in: query
        name: search
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Data found successfully
          schema:
            additionalProperties: true
            type: object
        "404":
          description: 'Not Found: No data found'
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      summary: Get a list of approval userss
      tags:
      - Approval User
    post:
      consumes:
      - application/json
      description: Create a new approval users with the provided details
      parameters:
      - description: Approval User details
        in: body
        name: approvalUser
        required: true
        schema:
          $ref: '#/definitions/model.MstApprovalUser'
      produces:
      - application/json
      responses:
        "201":
          description: Approval User created successfully
          schema:
            additionalProperties: true
            type: object
        "400":
          description: 'Bad Request: Invalid input'
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      summary: Create a new approval users
      tags:
      - Approval User
  /admin/approval-user/{id}:
    delete:
      description: Delete a approval users by its ID
      parameters:
      - description: Approval User ID
        in: path
        name: id
        required: true
        type: integer
      responses:
        "200":
          description: Approval User deleted successfully
          schema:
            additionalProperties: true
            type: object
        "400":
          description: 'Bad Request: Invalid ID'
          schema:
            additionalProperties: true
            type: object
        "404":
          description: 'Not Found: ApprovalUser not found'
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      summary: Delete a approval users
      tags:
      - Approval User
...
  /admin/master/employee:
    get:
      consumes:
      - application/json
      description: Fetch a paginated list of employees with optional search parameters
      parameters:
      - description: Page number (default is 1)
        in: query
        name: page
        type: integer
      - description: Number of rows per page (default is 20)
        in: query
        name: rows
        type: integer
      - description: Search term for employee name or number
        in: query
        name: search
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Data found successfully
          schema:
            additionalProperties: true
            type: object
        "400":
          description: 'Bad Request: Invalid query parameters'
          schema:
            additionalProperties: true
            type: object
        "404":
          description: 'Not Found: No data found'
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      summary: Retrieve Employee List
      tags:
      - Employee
  /admin/master/employee/{number}:
    get:
      consumes:
      - application/json
      description: Fetch the details of a specific employee using their employee number
      parameters:
      - description: Employee number
        in: path
        name: number
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Employee found successfully
          schema:
            additionalProperties: true
            type: object
        "400":
          description: 'Bad Request: Invalid employee number'
          schema:
            additionalProperties: true
            type: object
        "404":
          description: 'Not Found: Employee not found'
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      summary: Retrieve Employee by Number
      tags:
      - Employee
  /admin/master/employee/sync:
    get:
      consumes:
      - application/json
      description: Sync employee data from the external HRIS API and update the local
        database
      produces:
      - application/json
      responses:
        "200":
          description: Data synchronization successful
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      summary: Synchronize Employee Data
      tags:
      - Employee
...
swagger: "2.0"